finished adding permissions for different actions, users can't change permissions they don't have, there are checks on both the server and the client

+++ server/routes/api/bugs.js
@@ -18,6 +18,59 @@ router.get('/:projectId', (req, res) => {
 router.post('/:projectId', auth, (req, res) => {
   const bug = req.body;
   console.log("user object:", req.user)
+  db.one('SELECT * from users WHERE id = '+req.user.id)
+  .then(user => {
+    console.log("user:", user)
+    let values = [
+      bug.assigned_to,
+      bug.bug_name,
+      user.name,
+      user.id,
+      bug.deadline,
+      bug.hours_worked,
+      bug.percent_complete,
+      bug.severity,
+      bug.status,
+      bug.summary,
+      bug.time_estimate,
+      bug.version,
+    ];
+    var insertString = 'INSERT INTO bugs_'+req.params.projectId+' (assigned_to, bug_name, created_by, created_by_id, deadline, hours_worked, percent_complete, severity, status, summary, time_estimate, version, created_on) '
+      +'VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, current_timestamp) RETURNING *;'
+  
+    db.one('SELECT * from projects_users WHERE project_id = $1 AND user_id = $2', [req.params.projectId, req.user.id])
+    .then(current_project_user => {
+      if (current_project_user.add_bugs) {
+        console.log("project user:", current_project_user)
+  
+        console.log("insertString:", insertString)
+        console.log("values:", values)
+        db.one(insertString, values)
+        .then( newBug => {
+          console.log("newbug:",newBug)
+          res.json(newBug);
+        })
+        .catch(error => {
+          console.log(error);
+        })
+      } else {
+        return res.status(401).json({msg: "Current user doesn't have required permission"})
+      }
+    })
+    .catch(error => {
+      console.log(error)
+      return res.status(401).json({msg: "User is not a member of resource"})
+    })  
+  })
+  .catch (error => {
+    console.log(error)
+    return res.status(401).json({msg: "User is not logged in or unknown to system"})
+  })
+});
+
+router.patch('/:projectId', auth, (req, res) => {
+  const bug = req.body;
+  console.log("user object:", req.user)
   let values = [
     bug.assigned_to,
     bug.bug_name,
@@ -30,54 +83,69 @@ router.post('/:projectId', auth, (req, res) => {
     bug.summary,
     bug.time_estimate,
     bug.version,
+    bug.id
   ];
-  var insertString = 'INSERT INTO bugs_'+req.params.projectId+' (assigned_to, bug_name, created_by, deadline, hours_worked, percent_complete, severity, status, summary, time_estimate, version, created_on) '
-    +'VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, current_timestamp) RETURNING *;'
-  
-  if (bug.id !== "null"){
-    insertString = 'INSERT INTO bugs_'+req.params.projectId+' (assigned_to, bug_name, created_by, deadline, hours_worked, percent_complete, severity, status, summary, time_estimate, version, id, created_on) '
-      +'VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, current_timestamp) '
+  var insertString = 'INSERT INTO bugs_'+req.params.projectId+' (assigned_to, bug_name, deadline, hours_worked, percent_complete, severity, status, summary, time_estimate, version, id, created_on) '
+      +'VALUES ($1, $2, $4, $5, $6, $7, $8, $9, $10, $11, $12, current_timestamp) '
       +'ON CONFLICT (id) DO UPDATE '
-      +'SET assigned_to = $1, bug_name = $2, created_by = $3, deadline = $4, hours_worked = $5, percent_complete = $6, severity = $7, status = $8, summary = $9, time_estimate = $10, version = $11 RETURNING *;'
-    values = [...values, bug.id]
-  }
+      +'SET assigned_to = $1, bug_name = $2, deadline = $4, hours_worked = $5, percent_complete = $6, severity = $7, status = $8, summary = $9, time_estimate = $10, version = $11 RETURNING *;'
 
   db.one('SELECT * from projects_users WHERE project_id = $1 AND user_id = $2', [req.params.projectId, req.user.id])
-  .then(project_user => {
-    console.log("project user:", project_user)
+  .then(current_project_user => {
+    if (current_project_user.edit_bugs) {
+      console.log("project user:", current_project_user)
 
-    console.log("insertString:", insertString)
-    console.log("values:", values)
-    db.one(insertString, values)
-    .then( newBug => {
-      console.log("newbug:",newBug)
-      res.json(newBug);
-    })
-    .catch(error => {
-      console.log(error);
-    })
+      console.log("insertString:", insertString)
+      console.log("values:", values)
+      updateBugInDatabase(insertString, values, res)
+    } else if (current_project_user.edit_own_bugs) {
+      db.one('SELECT * from bugs_'+req.params.projectId+' WHERE created_by_id = $1 AND id = $2', [req.user.id, bug.id])
+      .then(bug => {
+        updateBugInDatabase(insertString, values, res)
+      }).catch(error => {
+        console.log(error)
+        return res.status(401).json({msg: "Current user didn't create resource and doesn't have required permission"})
+      })
+    } else {
+      return res.status(401).json({msg: "Current user doesn't have required permission"})
+    }
   })
   .catch(error => {
     console.log(error)
-    return res.status(401).json({msg: "User does not own resource"})
+    return res.status(401).json({msg: "User is not a member of resource"})
   })
 });
 
+function updateBugInDatabase(insertString, values, res){
+  db.one(insertString, values)
+  .then( newBug => {
+    console.log("newbug:",newBug)
+    res.json(newBug);
+  })
+  .catch(error => {
+    console.log(error);
+    return res.status(500).json({msg: "Error inserting bug into database"})
+  })
+}
 
 router.delete('/:id/:projectId', auth, (req, res) => {
   console.log("reached delete for bug api")
   console.log("user object:", req.user)
   db.one('SELECT * from projects_users WHERE project_id = $1 AND user_id = $2', [req.params.projectId, req.user.id])
-  .then(project_user => {
-    console.log("project user:", project_user)
-    db.none('DELETE from bugs_'+req.params.projectId+' WHERE id = '+req.params.id)
-    .then(t =>{
-      res.send(`Deleted bug with id=${req.params.id}`);
-    }).catch(error => res.status(404).send('bug id not found'))
+  .then(current_project_user => {
+    if (current_project_user.remove_bugs) {
+      console.log("project user:", current_project_user)
+      db.none('DELETE from bugs_'+req.params.projectId+' WHERE id = '+req.params.id)
+      .then(t =>{
+        res.send(`Deleted bug with id=${req.params.id}`);
+      }).catch(error => res.status(404).send('bug id not found'))
+    }  else {
+      return res.status(401).json({msg: "Current user doesn't have required permission"})
+    }
   })
   .catch(error => {
     console.log(error)
-    return res.status(401).json({msg: "User does not own resource"})
+    return res.status(401).json({msg: "User is not a member of resource"})
   })
 })
