+++ server/routes/api/projects.js
@@ -52,13 +52,13 @@ router.post('/', auth, (req, res) => {
     values
   ).then(newProject => {
     console.log(newProject)
-    db.none('INSERT INTO projects_users (project_id, user_id, created_on) '
-    +'VALUES ($1, $2, current_timestamp)',
-    [newProject.id, req.user.id]
+    db.none('INSERT INTO projects_users (project_id, user_id, created_by_id, delete_project, edit_project_page, add_users, remove_users, edit_users, edit_own_users, add_bugs, remove_bugs, edit_bugs, edit_own_bugs, created_on) '
+    +'VALUES ($1, $2, $3, true, true, true, true, true, true, true, true, true, true, current_timestamp)',
+    [newProject.id, req.user.id, req.user.id]
     ).catch(error => {
       console.log(error);
     })
-    db.none('CREATE TABLE IF NOT EXISTS bugs_$1 (id SERIAL PRIMARY KEY,  assigned_to VARCHAR(100), bug_name VARCHAR(100),  created_by VARCHAR(100),  deadline VARCHAR(100),  hours_worked VARCHAR(100),  percent_complete VARCHAR(100),  severity VARCHAR(100),  status VARCHAR(100),  summary VARCHAR(100),  time_estimate VARCHAR(100),  version VARCHAR(100), created_on DATE NOT NULL);', newProject.id);
+    db.none('CREATE TABLE IF NOT EXISTS bugs_$1 (id SERIAL PRIMARY KEY,  assigned_to VARCHAR(100), bug_name VARCHAR(100),  created_by VARCHAR(100), created_by_id SERIAL REFERENCES users (id), deadline VARCHAR(100),  hours_worked VARCHAR(100),  percent_complete VARCHAR(100),  severity VARCHAR(100),  status VARCHAR(100),  summary VARCHAR(100),  time_estimate VARCHAR(100),  version VARCHAR(100), created_on DATE NOT NULL);', newProject.id);
     res.json(newProject);
   })
   .catch(error => {
@@ -67,14 +67,25 @@ router.post('/', auth, (req, res) => {
 });
 
 router.delete('/:id', auth, (req, res) => {
-  db.any('DELETE from projects_users WHERE project_id = '+req.params.id+'RETURNING id')
+  db.one('SELECT * from projects_users WHERE project_id = $1 AND user_id = $2', [req.params.id, req.user.id])
+  .then(current_project_user => {
+    if (current_project_user.delete_project) {
+      db.any('DELETE from projects_users WHERE project_id = '+req.params.id+'RETURNING id')
 
-  db.none('DROP TABLE bugs_' + req.params.id)
+      db.none('DROP TABLE bugs_' + req.params.id)
 
-  db.none('DELETE from projects WHERE id = '+req.params.id)
-  .then(t =>{
-    res.send(`Deleted project with id=${req.params.id}`);
-  }).catch(error => res.status(400).send('project id not found'))
+      db.none('DELETE from projects WHERE id = '+req.params.id)
+      .then(t =>{
+        res.send(`Deleted project with id=${req.params.id}`);
+      }).catch(error => res.status(400).send('project id not found'))
+    } else {
+      return res.status(401).json({msg: "Current user doesn't have required permission"})
+    }
+  })
+  .catch(error => {
+    console.log(error)
+    return res.status(401).json({msg: "User is not a member of resource"})
+  })
 })
 
 module.exports = router
